requirements.txt
python-docx==1.1.2
opencc-python-reimplemented==0.1.7
openpyxl==3.1.5
PyYAML==6.0.2
pytest==8.3.5
regex==2024.11.6
PySide6==6.8.1
